CMS-14107 Use Maven CI friendly versions and use the revision property to manage the relesae version.
All project versions will be replaced with ${revision}, the revision property to be set in .mvn/maven.config.
The flatten plugin will expand the revision property to the project version on process-resources.
CI friendly versions are supported since Maven 3.5.0 so the enforced minimum Maven version is set to this as minimum required version.