SPASDK-2 Add exported function "getContainerItemContent"
The ContainerItem retrieves a model from the backend that, in case of a
valid x-component, contains a "content" property which references a
content item in the page scope.

The content reference is retrieved with the new method
getContentReference() on the ContainerItem interface.

To simplify the lookup of the container item content for an end-
developer, a new exported function is introduced named
getContainerItemContent(). It requires a ContainerItem and a Page
instance as arguments, and will return the data stored in the page
scope.

For version 0.9 we don't support a content reference on a
container-item and will always return undefined.