SPASDK-109 Describe solution to make persist the preview related data when navigated to the page without SPA SDK
If you are navigating between pages that have and those that don't have a SDK instance you will need to persist the preview data.
The easiest way to achieve this result is by making use of the cookie storage.
Parse a URl query string and extract preview related data as `token` and `server-id` and save them to the cookies.
If URL does not have query params than read cookies values and provide preview data as configuration to the SPA SDK `initialize` method.

Important notes
* The preview site should be served on a separate domain compared to the live site to avoid saving preview related cookies for the live site.
* In case, when the first page which is loading in preview doesn't use SPA SDK you should parse query parameters and save preview related data to the cookies by yourself.