SPASDK-119 Improve heroku deploy speed by using lerna filter flags
Lerna can filter the packages it is running commands on via its build in
filter flags. In this case we can use the `--scope` flag and the
`--include-dependencies` flag to only build our to-be-deployed app and
its dependencies. E.g.: the example react app, the react sdk and the spa
sdk.