SPASDK-126 Provide preparation of HTML content as page method
The sanitizing and rewriting links of HTML content is a common task.

Since we have all the information available in the page model anyway we
now provide a method to 'prepare' the HTML string with a single async
call. We have also provided a react hook for the React SDK for ease of
use.

Its an async process because under the hood a dynamic import of third
libraries is done such that those do not have to be included in the main
bundle.