SPASDK-122 Remove default apiVersion header value from PDA request
The apiVersion is 1.0 by default for SaaS and PaaS 15+, the header is
not even checked in those brxm versions. Therefore its only necessary to
send it along if you are using brxm 14.x and want to use apiVersion 1.0
instead of the default of 0.9. Adding the apiVersion header causes a
preflight request to be send which adds additional loading time to the
page so therefore we opt to remove it by default and require 14.x
customers to explicitly add it if they need to use PDA version 1.0.