SPASDK-148 Set cms update listener after loading the cms module
The spa module was managing the listener for the cms update event,
causing it to be set when the cms module had not loaded yet. Reversing
the control now lets the cms module determine when to call the spa
service update function to avoid this problem. It also just made more
sense to have the cms service control the cms update event call.