SPASDK-149 Clarify event origins in classes and modules
There are 2 event sets and eventbusses in place in the SDK. One that
communicates with the CMS UI (Experience Manager) and sets up the
preview integration and events, and then there is the Page events that
manage communication of the page ready event (and potentially others in
the future).

The Page event bus gets created *per page* where as the Cms event bus is
only created once.

The types/interfaces were named exacly the same leading to renaming
during imports and general confusion when reading the code as it not
immediately clear what event / eventbus was being injected/used.

Now the event and eventbus and related variables are named relative to
their origins, so PageEvents and CmsEvents for example, to clarify the
code.