SPASDK-149 Initialize now always returns a Promise<Page>
Initialized had the option to return a Page object directly,
synchronously. Unfortunately this had the side effect that some
initalization logic in the SDK was running in the wrong order. Also
quite frankly it was confusing to have this duality of synchronous and
asynchronous behavior in the same function.

This is a breaking change wrt the API, and is in line with the other
breaking changes in v18 where due to dynamic loading behavior we now use
asynchronous behavior in multiple places where we first had synchronous
behavior.