SPASDK-149 Revert "SPASDK-148 Set cms update listener after loading the cms module"
This reverts commit 57192c17

Read description in reverted commit first.
Turns out the SPA service can not be used in the cms service as the Cms
service is only instantiated once and the Spa service is instantiated
per page. Calling onCmsUpdate on the spa service would call it using an
spa service instance for another page after first navigation. Hence you
need the event bus to communicate to the newly instantiated spa
services.