SPASDK-149 Retrieve spa service from correct container
The spa service was injected into the cms service when the cms service
was constructed but the CMS service is a singleton, meaning that by the
time a user has navigated to another page and a new spa service was
constructed the cms service was still using an old spa service of which
the page had already been destroyed.

Passing in the scope into the cms service initialize method feels a bit
hacky to be honest. It may be better to convert the cms module back to a
sync container module and create providers for all its bindings instead.

(cherry picked from commit 8ba15d47bbc6fa615978bd3611c6896e6fb795ac)