SPASDK-176 Force rerender of meta tag when component changes
The component ID should be unique and should tell us whether the
component tree has updated. For currently unknown reason the meta html
comments were not being rerendered for certain components when the page
object changed, for example after navigation. Its likely to do with the
fact that the adding of the comments to the dom happens outside Vue code
and might not trigger the vue diff algorithm.