SPASDK-174 Only load CmsModule once
Previously it was possible to accidentally load the CmsModule
multiple times by calling the `initialize` method multiple
times in a row without waiting for the promise to load the CmsModule
asynchronously to be resolved, for example on a fast navigation. These
changes make sure that the CmsModule is only loaded once.