SPASDK-181 Revert async initialize and lazy load of preview code
When using SSR async behavior during the rendering of the component tree
is not executed in Vue, Nuxt.js, React, and Next.js because their
server-side rendering processes are synchronous by design. This means
that when components are rendered on the server, these frameworks will
not wait for async operations (e.g., Promises) to resolve before
generating the HTML. Therefore we can not rely on async initialize even
when no actual http request is done, and therefore we can not lazy load
preview code.