SPASDK-181 Remove manual changeDetection and its test
Removing the manual change detection seems to solve the rendering
problem with SSR, but the tests fails because the pure test module
does not 'see' the changed state after resolving the page model. It seem
like a smell, but at the same time there is no other indication that
things are not working in the real world.