SPASDK-181 Call sync on original page object in Vue 3 SDK
Vue 3 reactivity makes the page object deeply reactive. One of the side effects is that the object and its nested objects are not referencing the original objects. Because Emittery registers the Typed event emitters by instance calling the sync method on the reactive page object caused an error due the passed instance not being the registered one.