SPASDK-182 Remove the `sanitize` method from the Page object
The sanitize function is build into the SDK but its really a direct
method call to the sanitize-html third party library with 1 important
addition and that is the exclusion of the `data-type` attribute on <a>
tags from sanitizing. Other than that detail sanitizing the html is not
really part of the SDK concerns but rather the consumer SPA's concerns.

When creating an SPA that renders any rich content its to be expected
consumers of the SDK already have some sort of sanitizer library in
place. sanitize-html (or rather its dependencies) takes a significant
part of the bundle size, about 40%. After discussions we suggest to
remove the sanitize functionality from the SDK, and provide guidance in
the README instead.