SPASDK-191 Add link to typedoc
Tried integrating the typedoc directly into the astro build system but
failed due to incompatibilities between the typedoc output and astro
build requirements.

Considering that we will also split up the repo's in the future I've
just added a link to the to-be-deployed typedoc website as an external
page link.