SPASDK-231 Fix 500 issue in Nuxt example app
The Nuxt example app had a 500 Internal Server Error due to
mismatch in dependencies. Updating the dependencies to the latest
versions resolved the issue in prod mode. To address the editing
issue in the channel manager in dev mode, Nuxt config file
was updated to ensure that the emitter is not transpiled in the
build in dev mode